+++ packages/web3-shared/evm/src/helpers/createJsonRpcPayload.ts
@@ -1,4 +1,4 @@
-import type { JsonRpcPayload, JsonRpcResponse } from 'web3-core-helpers'
+import type { JsonRpcPayload } from 'web3-core-helpers'
 import type { RequestArguments } from '../types/index.js'
 
 export function createJsonRpcPayload(id: number, requestArguments: RequestArguments): JsonRpcPayload {
@@ -7,14 +7,5 @@ export function createJsonRpcPayload(id: number, requestArguments: RequestArgume
         id,
         method: requestArguments.method,
         params: requestArguments.params,
-    }
-}
-
-export function createJsonRpcResponse<T>(id: number, result?: T, error?: Error): JsonRpcResponse {
-    return {
-        jsonrpc: '2.0',
-        id,
-        result,
-        error,
     }
 }
