+++ packages/web3-shared/evm/src/helpers/createJsonRpcResponse.ts
@@ -1,0 +1,10 @@
+import type { JsonRpcResponse } from 'web3-core-helpers'
+
+export function createJsonRpcResponse<T>(id: number, result?: T, error?: Error): JsonRpcResponse {
+    return {
+        jsonrpc: '2.0',
+        id,
+        result,
+        error,
+    }
+}
