+++ packages/web3-shared/evm/src/helpers/encodeTypedData.ts
@@ -1,13 +1,9 @@
 import { type TypedDataDomain, utils, type TypedDataField } from 'ethers'
 
-export const splitSignature = (signature: string) => {
-    return utils.splitSignature(signature)
-}
-
-export const encodeTypedData = (
+export function encodeTypedData(
     domain: TypedDataDomain,
     types: Record<string, TypedDataField[]>,
     value: Record<string, any>,
-) => {
+) {
     return utils._TypedDataEncoder.getPayload(domain, types, value)
 }
