+++ packages/web3-shared/evm/src/helpers/isLens.ts
@@ -1,0 +1,5 @@
+export function isLens(name?: string) {
+    if (!name) return false
+    name = name.toLowerCase()
+    return name.endsWith('.lens') || name === 'lensprotocol' || name === '@lensprotocol'
+}
