+++ packages/web3-shared/evm/src/helpers/isLensFollower.ts
@@ -1,0 +1,4 @@
+export function isLensFollower(name: string) {
+    // vitalik.lens-Follower, lensprotocol-Follower V2
+    return name.includes('.lens-Follower') || name.includes('lensprotocol-Follower') || name.endsWith("'s follower NFT")
+}
