+++ packages/web3-shared/evm/src/helpers/resolveNonFungibleTokenIdFromEnsDomain.ts
@@ -1,0 +1,7 @@
+import { BigNumber } from 'bignumber.js'
+import { keccak256 } from '@ethersproject/keccak256'
+import { toUtf8Bytes } from '@ethersproject/strings'
+
+export function resolveNonFungibleTokenIdFromEnsDomain(domain: string): string {
+    return new BigNumber(keccak256(toUtf8Bytes(domain.replace(/\.\w+$/, '')))).toFixed()
+}
