+++ packages/web3-shared/evm/src/helpers/splitSignature.ts
@@ -1,0 +1,5 @@
+import { utils } from 'ethers'
+
+export function splitSignature(signature: string) {
+    return utils.splitSignature(signature)
+}
