+++ packages/web3-shared/evm/src/index.ts
@@ -1,21 +1,32 @@
-export * from './types/index.js'
 export * from './constants/index.js'
-export * from './libs/index.js'
-export * from './helpers/address.js'
-export * from './helpers/defaults.js'
-export * from './helpers/token.js'
-export * from './helpers/resolver.js'
-export * from './helpers/formatter.js'
-export * from './helpers/rpc.js'
 export * from './helpers/addGasMargin.js'
+export * from './helpers/address.js'
 export * from './helpers/createContract.js'
-export * from './helpers/decodeOutputString.js'
-export * from './helpers/parseStringOrBytes32.js'
-export * from './helpers/parseChainId.js'
-export * from './helpers/isValidDomain.js'
-export * from './helpers/isValidChainId.js'
+export * from './helpers/createJsonRpcPayload.js'
+export * from './helpers/createJsonRpcResponse.js'
 export * from './helpers/decodeEvents.js'
-export * from './helpers/getTransactionStatusType.js'
+export * from './helpers/decodeOutputString.js'
+export * from './helpers/defaults.js'
+export * from './helpers/encodeTypedData.js'
+export * from './helpers/formatter.js'
+export * from './helpers/getAverageBlockDelay.js'
+export * from './helpers/getContractOwnerDomain.js'
 export * from './helpers/getTransactionSignature.js'
+export * from './helpers/getTransactionStatusType.js'
+export * from './helpers/isLens.js'
+export * from './helpers/isLensCollect.js'
+export * from './helpers/isLensComment.js'
+export * from './helpers/isLensFollower.js'
+export * from './helpers/isLensPost.js'
+export * from './helpers/isValidChainId.js'
+export * from './helpers/isValidDomain.js'
+export * from './helpers/parseChainId.js'
+export * from './helpers/parseStringOrBytes32.js'
+export * from './helpers/resolveImageURL.js'
+export * from './helpers/resolveNonFungibleTokenIdFromEnsDomain.js'
+export * from './helpers/resolver.js'
 export * from './helpers/safeNonPayableTransactionCall.js'
-export * from './helpers/utils.js'
+export * from './helpers/splitSignature.js'
+export * from './helpers/token.js'
+export * from './libs/index.js'
+export * from './types/index.js'
